requirements.txt
attrs==25.3.0
certifi==2025.7.14
charset-normalizer==3.4.2
idna==3.10
isodate==0.7.2
lxml==6.0.0
platformdirs==4.3.8
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.4
requests-file==2.1.0
requests-toolbelt==1.0.0
urllib3==2.5.0
xmltodict==0.14.2
zeep==4.3.1
